# be/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreatePermissionRequest:
    properties:
      code:
        example: data.export
        type: string
      description:
        example: Can export data to CSV/Excel
        type: string
      module:
        example: data
        type: string
      name:
        example: Export Data
        type: string
    required:
    - code
    - module
    - name
    type: object
  models.CreateRoleRequest:
    properties:
      description:
        example: Editor role with content permissions
        type: string
      name:
        example: editor
        maxLength: 100
        minLength: 2
        type: string
      permission_ids:
        example:
        - a0000000-0000-0000-0000-000000000001
        items:
          type: string
        type: array
    required:
    - name
    type: object
...
  models.Permission:
    properties:
      code:
        example: user.view
        type: string
      created_at:
        type: string
      description:
        example: Can view user list and detail
        type: string
      id:
        example: a0000000-0000-0000-0000-000000000001
        type: string
      module:
        example: user
        type: string
      name:
        example: View Users
        type: string
      updated_at:
        type: string
    type: object
...
  models.Role:
    properties:
      created_at:
        type: string
      description:
        example: Administrator with management access
        type: string
      id:
        example: 10000000-0000-0000-0000-000000000002
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: admin
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
    type: object
...
  models.UpdatePermissionRequest:
    properties:
      code:
        example: data.export
        type: string
      description:
        example: Updated description
        type: string
      module:
        example: data
        type: string
      name:
        example: Export Data Updated
        type: string
    type: object
  models.UpdateRoleRequest:
    properties:
      description:
        example: Updated description
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: editor
        maxLength: 100
        minLength: 2
        type: string
      permission_ids:
        items:
          type: string
        type: array
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Meta'
      success:
        type: boolean
    type: object
  utils.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API dengan autentikasi lengkap, Dynamic RBAC, dan sistem menu
    hierarki.
  title: 247 Golang API
  version: "1.0"
paths:
...
  /permissions:
    get:
      description: Mendapatkan daftar semua permission, bisa difilter berdasarkan
        module. Memerlukan permission permission.view
      parameters:
      - description: 'Filter berdasarkan module (contoh: user, role, menu)'
        in: query
        name: module
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Permission'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Daftar semua permission
      tags:
      - Permissions
    post:
      consumes:
      - application/json
      description: Membuat permission baru. Memerlukan permission permission.create
      parameters:
      - description: Data permission baru
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Permission'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Buat permission baru
      tags:
      - Permissions
  /permissions/{id}:
    delete:
      description: Menghapus permission berdasarkan ID. Memerlukan permission permission.delete
      parameters:
      - description: Permission ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Hapus permission
      tags:
      - Permissions
    get:
      description: Mendapatkan detail permission berdasarkan ID. Memerlukan permission
        permission.view
      parameters:
      - description: Permission ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Permission'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Detail permission
      tags:
      - Permissions
    put:
      consumes:
      - application/json
      description: Mengupdate permission berdasarkan ID. Memerlukan permission permission.update
      parameters:
      - description: Permission ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Data update permission
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Permission'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update permission
      tags:
      - Permissions
  /roles:
    get:
      description: Mendapatkan daftar semua role beserta permission-nya. Memerlukan
        permission role.view
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Role'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Daftar semua role
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Membuat role baru dan assign permission. Memerlukan permission
        role.create
      parameters:
      - description: Data role baru
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Buat role baru
      tags:
      - Roles
  /roles/{id}:
    delete:
      description: Menghapus role berdasarkan ID. Memerlukan permission role.delete
      parameters:
      - description: Role ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Hapus role
      tags:
      - Roles
    get:
      description: Mendapatkan detail role beserta permission-nya. Memerlukan permission
        role.view
      parameters:
      - description: Role ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Detail role
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Mengupdate role dan assign ulang permission. Memerlukan permission
        role.update
      parameters:
      - description: Role ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Data update role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - Roles
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"